requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
python-doctr[cpu]
PyMuPDF
pillow
pytest
transformers>=4.21.0
torch>=1.12.0
datasets
scikit-learn
numpy>=1.21.0
pandas
celery[redis]
redis